requirements.txt
anyio==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
mutagen==1.47.0
protobuf==3.20.3
pydub==0.25.1
python-slugify==8.0.4
python-telegram-bot==20.8
pytube==15.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
text-unidecode==1.3
typing_extensions==4.9.0
urllib3==2.2.1
youtube-transcript-api==0.6.2
